Basic refactoring to get it building on linux
