--- GeomAlgoLib.sln
@@ -1,34 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29215.179
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "geom_algo_lib", "geom_algo_lib\geom_algo_lib.vcxproj", "{6D7FAB94-4E85-4EBD-8DF4-9795F753002D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RhinoInterface", "RhinoInterface\RhinoInterface.csproj", "{F04EDC90-2569-4052-B8CA-0C79E4853053}"
-	ProjectSection(ProjectDependencies) = postProject
-		{6D7FAB94-4E85-4EBD-8DF4-9795F753002D} = {6D7FAB94-4E85-4EBD-8DF4-9795F753002D}
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Release|x64 = Release|x64
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6D7FAB94-4E85-4EBD-8DF4-9795F753002D}.Debug|x64.ActiveCfg = Debug|x64
-		{6D7FAB94-4E85-4EBD-8DF4-9795F753002D}.Debug|x64.Build.0 = Debug|x64
-		{6D7FAB94-4E85-4EBD-8DF4-9795F753002D}.Release|x64.ActiveCfg = Release|x64
-		{6D7FAB94-4E85-4EBD-8DF4-9795F753002D}.Release|x64.Build.0 = Release|x64
-		{F04EDC90-2569-4052-B8CA-0C79E4853053}.Debug|x64.ActiveCfg = Debug|x64
-		{F04EDC90-2569-4052-B8CA-0C79E4853053}.Debug|x64.Build.0 = Debug|x64
-		{F04EDC90-2569-4052-B8CA-0C79E4853053}.Release|x64.ActiveCfg = Release|x64
-		{F04EDC90-2569-4052-B8CA-0C79E4853053}.Release|x64.Build.0 = Release|x64
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A2157DC3-C2B0-4666-9EB1-D45C2A1CADB3}
-	EndGlobalSection
-EndGlobal
