--- RhinoInterface/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="RhinoCommon" version="6.30.20288.16411" targetFramework="net472" />
-</packages>
