--- geom_algo_lib/cpp.hint
@@ -1,4 +1,0 @@
-// Hint files help the Visual Studio IDE interpret Visual C++ identifiers
-// such as names of functions and macros.
-// For more information see https://go.microsoft.com/fwlink/?linkid=865984
-#define PINVOKE extern "C" __declspec(dllexport)
