--- geom_algo_lib/framework.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#define WIN32_LEAN_AND_MEAN             // Exclude rarely-used stuff from Windows headers
-// Windows Header Files
-#include <windows.h>
