--- geom_algo_lib/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="boost" version="1.72.0.0" targetFramework="native" />
-</packages>
